Remove unused methods, move GradesStore to instructor_task/models.py
